requirements.txt
pyTelegramBotAPI
qrcode
pillow
python-dotenv
flask
requests
